rename: `jwtUtil` auth.jwt로 이동 (#23)
