+++ src/main/java/team05/integrated_feed_backend/core/config/SecurityConfig.java
@@ -10,7 +10,7 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import team05.integrated_feed_backend.common.util.JwtUtil;
+import team05.integrated_feed_backend.module.auth.jwt.JwtUtil;
 import team05.integrated_feed_backend.module.auth.jwt.JwtAuthenticationFilter;
 
 @Configuration
