+++ src/main/java/team05/integrated_feed_backend/module/auth/jwt/JwtAuthenticationFilter.java
@@ -9,7 +9,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.web.filter.OncePerRequestFilter;
-import team05.integrated_feed_backend.common.util.JwtUtil;
 
 import java.io.IOException;
